wpa_supplicant_lib: Fix integer overflow in result_copy_to_buf
In the function result_copy_to_buf, the result of strlen, which
is a size_t, is assigned to a variable of type int. This can
cause the bounds check on the size of dst_buf to succeed even if
src is very large, resulting in a buffer overflow on os_memcpy.

This CL fixes the integer overflow to avoid buffer overflow on
os_memcpy.

CRs-Fixed: 3614895
Change-Id: Ie77c3e4c76e825e649a1c32b51e05de06cb8ae26