wifihal: Fix to fetch peer ctx on Wifi disable/enable for initiator
When WiFi or NAN is disabled and enabled, peer context is flushed and
peer comes up with a new NMI address. Hence create peer context with
required parameters for pairing verification request for initiator,
if it is not available.

PMK received from framework is added to the cache before sending
the verification request.

CRs-Fixed: 3622015
Change-Id: I0059322c6eb92ada199a0629d9692aa7fb2247d2