wifihal: Fix to fetch peer ctx on Wifi disable/enable for responder
When WiFi or NAN is disabled and enabled, peer context is flushed and
peer comes up with a new NMI address. Hence create peer context with
required parameters for pairing verification request for responder,
if it is not available. set the is_paired to true, to validate the
verification request.

PMK received from framework is added to the cache before processing
the verification request Indication received.

CRs-Fixed: 3622015
Change-Id: I42149b6aa43e15dbad3d27fb396d28f0e96f0d68